move actorregister to common
