+++ transformer/src/main/scala/uk/ac/wellcome/transformer/actors/DynamoCaseClassExtractorActor.scala
@@ -9,7 +9,7 @@ import uk.ac.wellcome.models.CalmDynamoRecord
 
 import javax.inject.Inject
 import uk.ac.wellcome.platform.transformer.modules.ActorRegistryModule
-import uk.ac.wellcome.platform.transformer.modules.ActorRegister
+import uk.ac.wellcome.models.ActorRegister
 
 import com.google.inject.name.Named
 
@@ -24,14 +24,11 @@ class DynamoCaseClassExtractorActor @Inject()(
       ScanamoFree.read[CalmDynamoRecord](record.value) match {
         case Right(o) => {
           info(s"Parsed DynamoDB record ${o}")
-
-          actorRegister.actors
-            .get("transformActor")
-            .map(_ ! o)
+          actorRegister.send("transformActor", o)
         }
         case Left(o) => {
           error(s"Unable to parse record ${o}")
-          // Send to dead letter queue or error
+          // TODO: Send to dead letter queue or error
         }
       }
     }
