+++ transformer/src/main/scala/uk/ac/wellcome/transformer/actors/KinesisDynamoRecordExtractorActor.scala
@@ -8,7 +8,7 @@ import com.twitter.inject.Logging
 import javax.inject.Inject
 import com.twitter.inject.TwitterModule
 import uk.ac.wellcome.platform.transformer.modules.ActorRegistryModule
-import uk.ac.wellcome.platform.transformer.modules.ActorRegister
+import uk.ac.wellcome.models.ActorRegister
 import com.google.inject.name.Named
 
 case class RecordMap(value: java.util.Map[String, AttributeValue])
@@ -27,9 +27,7 @@ class KinesisDynamoRecordExtractorActor @Inject()(
         .getNewImage()
 
       info(s"Received record ${keys}")
-      actorRegister.actors
-        .get("dynamoCaseClassExtractorActor")
-        .map(_ ! RecordMap(keys))
+      actorRegister.send("dynamoCaseClassExtractorActor", RecordMap(keys))
     }
     case event => error(s"Received unknown Kinesis event ${event}")
   }
