+++ transformer/src/main/scala/uk/ac/wellcome/transformer/actors/PublishableMessageRecordActor.scala
@@ -13,7 +13,7 @@ import uk.ac.wellcome.platform.transformer.modules.{
   ActorRegistryModule
 }
 
-import uk.ac.wellcome.platform.transformer.modules.ActorRegister
+import uk.ac.wellcome.models.ActorRegister
 
 import com.amazonaws.services.sns.AmazonSNS
 
@@ -42,13 +42,10 @@ class PublishableMessageRecordActor @Inject()(
           )
 
           info(s"Publishable message ${message}")
-
-          actorRegister.actors
-            .get("publisherActor")
-            .map(_ ! message)
+          actorRegister.send("publisherActor", message)
         }
         case Failure(e) => {
-          // Send to dead letter queue or just error
+          // TODO: Send to dead letter queue or just error
           error("Failed to convert into publishable message", e)
         }
       }
