+++ transformer/src/main/scala/uk/ac/wellcome/transformer/actors/PublisherActor.scala
@@ -3,7 +3,6 @@ package uk.ac.wellcome.platform.transformer.actors
 import akka.actor.Actor
 import com.twitter.inject.Logging
 
-import uk.ac.wellcome.platform.transformer.models._
 import uk.ac.wellcome.platform.transformer.modules.PublishableMessage
 
 import scala.util.Success
