+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/ActorRegistryModule.scala
@@ -7,6 +7,7 @@ import akka.actor.{ActorRef, ActorSystem, Props}
 
 import uk.ac.wellcome.platform.finatra.modules._
 import uk.ac.wellcome.platform.transformer.actors._
+import uk.ac.wellcome.models.ActorRegister
 
 import uk.ac.wellcome.utils.GuiceAkkaExtension
 import net.codingwell.scalaguice.ScalaModule
@@ -14,7 +15,6 @@ import net.codingwell.scalaguice.ScalaModule
 import akka.actor.Actor
 import com.google.inject.name.Names
 
-case class ActorRegister(actors: Map[String, ActorRef])
 
 object ActorRegistryModule extends TwitterModule {
