+++ transformer/src/main/scala/uk/ac/wellcome/transformer/modules/StreamRecordProcessorFactoryModule.scala
@@ -14,6 +14,7 @@ import com.twitter.inject.TwitterModule
 import akka.actor.ActorRef
 import uk.ac.wellcome.platform.finatra.modules._
 import uk.ac.wellcome.platform.transformer.modules._
+import uk.ac.wellcome.models.ActorRegister
 
 class StreamsRecordProcessor(
   client: AmazonDynamoDB,
